requirements.txt
boto3
requests
openai
load_dotenv
pymongo
bson
